api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
httpx==0.27.0
python-dotenv==1.0.1
